Multiple changes to field names and a few table names to reflect docstore database schema changes.
I commented out a bunch of Code in LU_BuildInstance's readSirsiFiles
method that was populating the callNumbersByCatalogKey and
callNumbersByItemNumber hashmaps, and the itemsByCatalogKey hashmap.
The only one still in use is itemsByID.  I commented out and removed
code from several other methods that were making unnecessary references
to the now defunct hashmaps.  This reduces the memory requirement for
running the loader a bit, though it still needs 4 gigs of RAM to keep
the itemsByID hash all in memory.

I added a Tokenizer class to help in parsing coverage dates.  At first I
was going to try to create a grammar for the free-text that was entered
into the 856 and 866 fields in the catalog, but it turns out that for
e-journals, SFX contains better data.  So, I used the Tokenizer to parse
that data and added several methods to test2.java to read the SFX data,
compare it to the catalog data, and parse Coverage information from the
SFX export.