Added Deliver_ItemStatus and Deliver_ItemType classes, which are contained in the ItemStatus and ItemType classes.  Basically the ItemStatus and ItemType data needs to be in two tables.  Adding these classes and putting them into the existing ones seemed like the easiest way to achieve this.
I added unique prefix IDs for bibs (wbo), holdings (who), and item (wio)
because in places the OLE code seems to assume those prefixes will be
there.

I added code to buildItemsData in LU_BuildInstance that assigns the
item's status code and name using the number of charges or the reserve
status if applicable.