Bib IDs are now preserved from Sirsi.  The bib_id column of ole_ds_bib_t is no longer generated.
MARC holdings records are no longer loaded into ole_ds_bib_t, since they
would cause a primary key conflict with the bib records they correspond
to.

I now give a default value to the shelving order of a holdings record,
and also compute one using the same code that OLE uses for Dewey and LoC
call numbers.  I haven't found a solution yet for SuDoc call numbers.

OCLC title control numbers are now added to records that are missing
them in an 035 field.  New and existing OCLC  and are formatted
according to a scheme that cataloging librarian Lisa McColl specified.
Exact duplicate 035 fields are eliminated.

The serials and issues are now read from a SQLite staging database
called olemigration, rather than a MySQL database.  This seems to have
improved the performance of the loader quite a bit.  The SQLite database
is in a file called olemigration.sqlite, and the path to it must be
given with "-Dmigrationdbdir=<path to file>" on the command line call to
the loader.  The migrationdbdir variable is then used in persistence.xml
for the SQLite database URL.  I changed pom.xml to bring in the JDBC
driver for SQLite, and added the org.hibernate.dialect.SQLiteDialect
class to my project.  I found that dialect code at
https://github.com/gwenn/sqlite-dialect/blob/master/src/main/java/org/hibernate/dialect/SQLiteDialect.java.
I also had to change the types of all the ID fields in the migration
database objects to be "int" rather than "Long" because of how the
SQLiteDialect works.