Serials receiving loader now creates FDoc RICE documents in RICE tables (krew_doc_hdr_t, krns_doc_hdr_t, krew_doc_hdr_cntnt_t) and links serials receiving control records to them.  The document type in krew_doc_hdr_t is just set to 999999 because the document type for serials receiving doesn't actually exist until OLE is running.  It's created in krew_doc_typ_t as the application is unpacked.  Thus, the build process starts tomcat, waits 2 minutes, and then updates all the documents with type 999999 to have the new document type.
The loader creates boundwiths now using a new export from Sirsi.

The location level string is now built into the FlatLocation class.

I gave the item damaged status and missing pieces fields of the Item
class default values of "N".

The Location class now links to the ole_locn_t table and all of its
fields.  Locations are read from the OLE database rather than from a
.csv file.

Sirsi export files are no longer read into hashmaps, but into the
olemigration SQLite database.  This allows the memory footprints of the
loader to be smaller, but it's slow if this SQLite database is on disk,
so I often end up putting the migration database into a ramdisk.

The LocationLevel class now links to the ole_locn_level_t table.

The formerID field is now added to the OLEHoldings class, and links to
the FORMER_HOLDINGS_ID field in the database.

Locations that have been renamed have a hashmap in LU_BuildInstance to
map the old names to new ones.

Items use the locations from the current location field, rather than the
home location field now.

Call numbers are normalized using CallNumUtils.java now, rather than the
DeweyCallNumber and LCCallNumber classes from SolrMarc.

LU_DBLoadInstances now takes a parameter about whether or not to stage
data in the migration database.  Since it is persistent, this only needs
to be done once for any given export, and the loader sometimes takes
multiple tries to succeed, so it can save significant time to not do
this each time.